+++ test/napkinsketch_book/cookbook_generated_test.clj
@@ -102,8 +102,8 @@
    :value
    (map
     (fn*
-     [p1__82045#]
-     (+ 100.0 (* 30.0 (Math/sin (* (double p1__82045#) 0.12)))))
+     [p1__85932#]
+     (+ 100.0 (* 30.0 (Math/sin (* (double p1__85932#) 0.12)))))
     (range 52))}))
 
 
@@ -433,7 +433,7 @@
      (and
       (pos? (:points s))
       (some
-       (fn* [p1__82046#] (= "virginica" p1__82046#))
+       (fn* [p1__85933#] (= "virginica" p1__85933#))
        (:texts s)))))
    v64_l287)))
 
@@ -487,8 +487,8 @@
  (->
   {:x (range 20),
    :y
-   (map (fn* [p1__82047#] (Math/sin (/ p1__82047# 3.0))) (range 20)),
-   :change (map (fn* [p1__82048#] (- p1__82048# 10)) (range 20))}
+   (map (fn* [p1__85934#] (Math/sin (/ p1__85934# 3.0))) (range 20)),
+   :change (map (fn* [p1__85935#] (- p1__85935# 10)) (range 20))}
   (sk/lay-point :x :y {:color :change})
   (sk/options
    {:color-scale :diverging,
@@ -605,8 +605,8 @@
    ys
    (map
     (fn*
-     [p1__82049#]
-     (+ (* 3 p1__82049#) 5 (* 2 (- (rng/drandom r) 0.5))))
+     [p1__85936#]
+     (+ (* 3 p1__85936#) 5 (* 2 (- (rng/drandom r) 0.5))))
     xs)]
   (->
    {:x xs, :y ys}